Misc:Update
Former-commit-id: 1810cc7075cca4ac5daccdbe7f8112b333aad6c9 [formerly 3f054672d4e110bcb02f7704c04855735249c0db]
Former-commit-id: 9161b13af4c421c3ab18edc783c7cc74a7475d68

+++ IoT/RC_car_control/main.py
@@ -386,9 +386,9 @@ class MyApp(QMainWindow, Ui_MainWindow):
     def handleSave(self):
         global init_data, left, center, right
         
-        left = int(self.ui.le_handle_left.text())
-        center = int(self.ui.le_handle_center.text())
-        right = int(self.ui.le_handle_right.text())
+        left = float(self.ui.le_handle_left.text())
+        center = float(self.ui.le_handle_center.text())
+        right = float(self.ui.le_handle_right.text())
         
         init_data['Handle_left'] = left
         init_data['Handle_center'] = center
